:truck: Move diff into own module in configuration subpackage
