+++ CPAC/utils/configuration/__init__.py
@@ -1,0 +1,24 @@
+"""C-PAC Configuration module
+
+Copyright (C) 2022  C-PAC Developers
+
+This file is part of C-PAC.
+
+C-PAC is free software: you can redistribute it and/or modify it under
+the terms of the GNU Lesser General Public License as published by the
+Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+C-PAC is distributed in the hope that it will be useful, but WITHOUT
+ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with C-PAC. If not, see <https://www.gnu.org/licenses/>."""
+from .configuration import Configuration, DEFAULT_PIPELINE_FILE, \
+                           Preconfiguration
+from . import configuration, diff
+
+__all__ = ['Configuration', 'configuration', 'DEFAULT_PIPELINE_FILE', 'diff',
+           'Preconfiguration']
