+++ CPAC/utils/configuration/configuration.py
@@ -1,4 +1,21 @@
-"""C-PAC Configuration class"""
+"""C-PAC Configuration class and related functions
+
+Copyright (C) 2022  C-PAC Developers
+
+This file is part of C-PAC.
+
+C-PAC is free software: you can redistribute it and/or modify it under
+the terms of the GNU Lesser General Public License as published by the
+Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+C-PAC is distributed in the hope that it will be useful, but WITHOUT
+ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with C-PAC. If not, see <https://www.gnu.org/licenses/>."""
 import re
 import os
 import warnings
@@ -8,7 +25,8 @@ from warnings import warn
 
 import yaml
 
-from CPAC.utils.utils import dct_diff, load_preconfig
+from CPAC.utils.utils import load_preconfig
+from .diff import dct_diff
 
 SPECIAL_REPLACEMENT_STRINGS = {r'${resolution_for_anat}',
                                r'${func_resolution}'}
@@ -19,7 +37,7 @@ DEFAULT_PIPELINE_FILE = '/cpac_resources/default_pipeline.yml'
 if not os.path.exists(DEFAULT_PIPELINE_FILE):
     CPAC_DIRECTORY = os.path.abspath(os.path.join(
         __file__,
-        *repeat(os.path.pardir, 3)))
+        *repeat(os.path.pardir, 4)))
     # package location
     DEFAULT_PIPELINE_FILE = os.path.join(
         CPAC_DIRECTORY,
@@ -180,7 +198,40 @@ class Configuration:
             self.key_type_error(key)
 
     def __sub__(self: 'Configuration', other: 'Configuration'):
-        '''Return the set difference between two Configurations'''
+        '''Return the set difference between two Configurations
+
+        Examples
+        --------
+        >>> diff = (Preconfiguration('fmriprep-options') - Configuration()) \
+        # doctest: +NORMALIZE_WHITESPACE
+        Loading the 'fmriprep-options' pre-configured pipeline.
+        >>> diff['pipeline_setup']['pipeline_name']
+        ('cpac_fmriprep-options', 'cpac-default-pipeline')
+        >>> diff['pipeline_setup']['pipeline_name'].s_value
+        'cpac_fmriprep-options'
+        >>> diff['pipeline_setup']['pipeline_name'].t_value
+        'cpac-default-pipeline'
+        >>> diff.s_value['pipeline_setup']['pipeline_name']
+        'cpac_fmriprep-options'
+        >>> diff.t_value['pipeline_setup']['pipeline_name']
+        'cpac-default-pipeline'
+        >>> diff['pipeline_setup']['pipeline_name'].left
+        'cpac_fmriprep-options'
+        >>> diff.left['pipeline_setup']['pipeline_name']
+        'cpac_fmriprep-options'
+        >>> diff['pipeline_setup']['pipeline_name'].minuend
+        'cpac_fmriprep-options'
+        >>> diff.minuend['pipeline_setup']['pipeline_name']
+        'cpac_fmriprep-options'
+        >>> diff['pipeline_setup']['pipeline_name'].right
+        'cpac-default-pipeline'
+        >>> diff.right['pipeline_setup']['pipeline_name']
+        'cpac-default-pipeline'
+        >>> diff['pipeline_setup']['pipeline_name'].subtrahend
+        'cpac-default-pipeline'
+        >>> diff.subtrahend['pipeline_setup']['pipeline_name']
+        'cpac-default-pipeline'
+        '''
         return(dct_diff(self.dict(), other.dict()))
 
     def dict(self):
