+++ CPAC/utils/yaml_template.py
@@ -1,10 +1,28 @@
+"""Copyright (C) 2022  C-PAC Developers
+
+This file is part of C-PAC.
+
+C-PAC is free software: you can redistribute it and/or modify it under
+the terms of the GNU Lesser General Public License as published by the
+Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+C-PAC is distributed in the hope that it will be useful, but WITHOUT
+ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with C-PAC. If not, see <https://www.gnu.org/licenses/>."""
 import os
 import re
+from datetime import datetime
 import yaml
 from click import BadParameter
-from datetime import datetime
+
 from CPAC.utils.configuration import Configuration, DEFAULT_PIPELINE_FILE
-from CPAC.utils.utils import dct_diff, diff_dict, load_preconfig, \
+from CPAC.utils.configuration.diff import dct_diff, diff_dict
+from CPAC.utils.utils import load_preconfig, \
                              lookup_nested_value, update_config_dict, \
                              update_pipeline_values_1_8
